+++ flutter_module/lib/kline/klinedatamanager.dart
